[&1.6] Issue 3783: 	Breakpoint mark fails to show up when Firebug is first detached in to a window.
When the Firebug window is detached, initializeUI is fired, which resets the textSize, which erases all of the sourceBoxes so they will be recreated. removeAllSourceBoxes does the work, but failed to also clear the selectedSourceBox. When the detached window came up, the selectedSourceBox was selected, but it was not in the list of sourceBoxes.

This bug should have been caught by FBTest openInNewWindow?

http://code.google.com/p/fbug/source/detail?r=8676